+++ cmd/vegawallet/commands/message_verify.go
@@ -11,6 +11,7 @@ import (
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/cli"
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/flags"
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/printer"
+	"code.vegaprotocol.io/vega/libs/jsonrpc"
 	"code.vegaprotocol.io/vega/wallet/api"
 
 	"github.com/spf13/cobra"
@@ -34,7 +35,7 @@ type VerifyMessageHandler func(api.AdminVerifyMessageParams) (api.AdminVerifyMes
 func NewCmdVerifyMessage(w io.Writer, rf *RootFlags) *cobra.Command {
 	h := func(params api.AdminVerifyMessageParams) (api.AdminVerifyMessageResult, error) {
 		verifyMessage := api.NewAdminVerifyMessage()
-		rawResult, errorDetails := verifyMessage.Handle(context.Background(), params)
+		rawResult, errorDetails := verifyMessage.Handle(context.Background(), params, jsonrpc.RequestMetadata{})
 		if errorDetails != nil {
 			return api.AdminVerifyMessageResult{}, errors.New(errorDetails.Data)
 		}