+++ cmd/vegawallet/commands/network_list.go
@@ -9,6 +9,7 @@ import (
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/cli"
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/flags"
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/printer"
+	"code.vegaprotocol.io/vega/libs/jsonrpc"
 	"code.vegaprotocol.io/vega/paths"
 	"code.vegaprotocol.io/vega/wallet/api"
 	networkStore "code.vegaprotocol.io/vega/wallet/network/store/v1"
@@ -39,7 +40,7 @@ func NewCmdListNetworks(w io.Writer, rf *RootFlags) *cobra.Command {
 		}
 
 		listWallet := api.NewAdminListNetworks(networkStore)
-		rawResult, errorDetails := listWallet.Handle(context.Background(), nil)
+		rawResult, errorDetails := listWallet.Handle(context.Background(), nil, jsonrpc.RequestMetadata{})
 		if errorDetails != nil {
 			return api.AdminListNetworksResult{}, errors.New(errorDetails.Data)
 		}