+++ cmd/vegawallet/commands/permissions_list.go
@@ -9,6 +9,7 @@ import (
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/cli"
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/flags"
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/printer"
+	"code.vegaprotocol.io/vega/libs/jsonrpc"
 	"code.vegaprotocol.io/vega/wallet/api"
 	"code.vegaprotocol.io/vega/wallet/wallets"
 
@@ -36,7 +37,7 @@ func NewCmdListPermissions(w io.Writer, rf *RootFlags) *cobra.Command {
 		}
 
 		listPermissions := api.NewAdminListPermissions(s)
-		rawResult, errDetails := listPermissions.Handle(context.Background(), params)
+		rawResult, errDetails := listPermissions.Handle(context.Background(), params, jsonrpc.RequestMetadata{})
 		if errDetails != nil {
 			return api.AdminListPermissionsResult{}, errors.New(errDetails.Data)
 		}