+++ cmd/vegawallet/commands/transaction_sign.go
@@ -11,6 +11,7 @@ import (
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/cli"
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/flags"
 	"code.vegaprotocol.io/vega/cmd/vegawallet/commands/printer"
+	"code.vegaprotocol.io/vega/libs/jsonrpc"
 	"code.vegaprotocol.io/vega/paths"
 	coreversion "code.vegaprotocol.io/vega/version"
 	"code.vegaprotocol.io/vega/wallet/api"
@@ -76,7 +77,7 @@ func NewCmdSignTransaction(w io.Writer, rf *RootFlags) *cobra.Command {
 			return walletnode.BuildRoundRobinSelectorWithRetryingNodes(log, hosts, retries)
 		})
 
-		rawResult, errDetails := signTx.Handle(context.Background(), params)
+		rawResult, errDetails := signTx.Handle(context.Background(), params, jsonrpc.RequestMetadata{})
 		if errDetails != nil {
 			return api.AdminSignTransactionResult{}, errors.New(errDetails.Data)
 		}