+++ wallet/api/admin_close_connections_to_hostname_test.go
@@ -56,7 +56,7 @@ func testAdminCloseConnectionsToHostnameWithInvalidParamsFails(t *testing.T) {
 	for _, tc := range tcs {
 		t.Run(tc.name, func(tt *testing.T) {
 			// given
-			ctx, _ := contextWithTraceID()
+			ctx := context.Background()
 
 			// setup
 			handler := newCloseConnectionsToHostnameHandler(tt)
@@ -242,7 +242,7 @@ type adminCloseConnectionsToHostnameHandler struct {
 func (h *adminCloseConnectionsToHostnameHandler) handle(t *testing.T, ctx context.Context, params interface{}) *jsonrpc.ErrorDetails {
 	t.Helper()
 
-	rawResult, err := h.Handle(ctx, params)
+	rawResult, err := h.Handle(ctx, params, jsonrpc.RequestMetadata{})
 	require.Empty(t, rawResult)
 	return err
 }