+++ wallet/api/admin_describe_network_test.go
@@ -50,7 +50,7 @@ func testDescribingNetworkWithInvalidParamsFails(t *testing.T) {
 	for _, tc := range tcs {
 		t.Run(tc.name, func(tt *testing.T) {
 			// given
-			ctx, _ := contextWithTraceID()
+			ctx := context.Background()
 
 			// setup
 			handler := newDescribeNetworkHandler(tt)
@@ -167,7 +167,7 @@ type describeNetworkHandler struct {
 func (h *describeNetworkHandler) handle(t *testing.T, ctx context.Context, params interface{}) (api.AdminDescribeNetworkResult, *jsonrpc.ErrorDetails) {
 	t.Helper()
 
-	rawResult, err := h.Handle(ctx, params)
+	rawResult, err := h.Handle(ctx, params, jsonrpc.RequestMetadata{})
 	if rawResult != nil {
 		result, ok := rawResult.(api.AdminDescribeNetworkResult)
 		if !ok {