+++ wallet/api/admin_list_keys_test.go
@@ -56,7 +56,7 @@ func testAdminListKeysWithInvalidParamsFails(t *testing.T) {
 	for _, tc := range tcs {
 		t.Run(tc.name, func(tt *testing.T) {
 			// given
-			ctx, _ := contextWithTraceID()
+			ctx := context.Background()
 
 			// setup
 			handler := newAdminListKeysHandler(tt)
@@ -179,7 +179,7 @@ type adminListKeysHandler struct {
 func (h *adminListKeysHandler) handle(t *testing.T, ctx context.Context, params interface{}) (api.AdminListKeysResult, *jsonrpc.ErrorDetails) {
 	t.Helper()
 
-	rawResult, err := h.Handle(ctx, params)
+	rawResult, err := h.Handle(ctx, params, jsonrpc.RequestMetadata{})
 	if rawResult != nil {
 		result, ok := rawResult.(api.AdminListKeysResult)
 		if !ok {