+++ wallet/api/admin_sign_message.go
@@ -26,7 +26,7 @@ type AdminSignMessage struct {
 	walletStore WalletStore
 }
 
-func (h *AdminSignMessage) Handle(ctx context.Context, rawParams jsonrpc.Params) (jsonrpc.Result, *jsonrpc.ErrorDetails) {
+func (h *AdminSignMessage) Handle(ctx context.Context, rawParams jsonrpc.Params, _ jsonrpc.RequestMetadata) (jsonrpc.Result, *jsonrpc.ErrorDetails) {
 	params, err := validateAdminSignMessageParams(rawParams)
 	if err != nil {
 		return nil, invalidParams(err)