+++ wallet/api/admin_stop_service_test.go
@@ -48,7 +48,7 @@ func testStoppingNetworkWithInvalidParamsFails(t *testing.T) {
 	for _, tc := range tcs {
 		t.Run(tc.name, func(tt *testing.T) {
 			// given
-			ctx, _ := contextWithTraceID()
+			ctx := context.Background()
 
 			// setup
 			handler := newStopServiceHandler(tt)
@@ -117,7 +117,7 @@ type stopNetworkHandler struct {
 func (h *stopNetworkHandler) handle(t *testing.T, ctx context.Context, params interface{}) *jsonrpc.ErrorDetails {
 	t.Helper()
 
-	rawResult, err := h.Handle(ctx, params)
+	rawResult, err := h.Handle(ctx, params, jsonrpc.RequestMetadata{})
 	assert.Nil(t, rawResult)
 	return err
 }