+++ wallet/api/client_connect_wallet.go
@@ -9,7 +9,6 @@ import (
 	"code.vegaprotocol.io/vega/wallet/api/session"
 	"code.vegaprotocol.io/vega/wallet/preferences"
 	"code.vegaprotocol.io/vega/wallet/wallet"
-	"github.com/mitchellh/mapstructure"
 )
 
 const WalletConnectionSuccessfullyEstablished = "The connection to the wallet has been successfully established."
@@ -18,10 +17,6 @@ type ClientConnectWallet struct {
 	walletStore WalletStore
 	interactor  Interactor
 	sessions    *session.Sessions
-}
-
-type ClientConnectWalletParams struct {
-	Hostname string `json:"hostname"`
 }
 
 type ClientConnectWalletResult struct {
@@ -40,43 +35,36 @@ type ClientConnectWalletResult struct {
 //
 // All sessions have to be initialized by using this handler. Otherwise, a call
 // to any other handlers will be rejected.
-func (h *ClientConnectWallet) Handle(ctx context.Context, rawParams jsonrpc.Params) (jsonrpc.Result, *jsonrpc.ErrorDetails) {
-	traceID := TraceIDFromContext(ctx)
-
-	params, err := validateConnectWalletParams(rawParams)
-	if err != nil {
-		return nil, invalidParams(err)
-	}
-
-	if err := h.interactor.NotifyInteractionSessionBegan(ctx, traceID); err != nil {
+func (h *ClientConnectWallet) Handle(ctx context.Context, _ jsonrpc.Params, metadata jsonrpc.RequestMetadata) (jsonrpc.Result, *jsonrpc.ErrorDetails) {
+	if err := h.interactor.NotifyInteractionSessionBegan(ctx, metadata.TraceID); err != nil {
 		return nil, internalError(err)
 	}
-	defer h.interactor.NotifyInteractionSessionEnded(ctx, traceID)
+	defer h.interactor.NotifyInteractionSessionEnded(ctx, metadata.TraceID)
 
 	availableWallets, err := h.walletStore.ListWallets(ctx)
 	if err != nil {
-		h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("could not list the available wallets: %w", err))
+		h.interactor.NotifyError(ctx, metadata.TraceID, InternalError, fmt.Errorf("could not list the available wallets: %w", err))
 		return nil, internalError(ErrCouldNotConnectToWallet)
 	}
 	if len(availableWallets) == 0 {
-		h.interactor.NotifyError(ctx, traceID, ApplicationError, ErrNoWalletToConnectTo)
+		h.interactor.NotifyError(ctx, metadata.TraceID, ApplicationError, ErrNoWalletToConnectTo)
 		return nil, applicationCancellationError(ErrApplicationCanceledTheRequest)
 	}
 
 	var approval preferences.ConnectionApproval
 	for {
-		rawApproval, err := h.interactor.RequestWalletConnectionReview(ctx, traceID, params.Hostname)
+		rawApproval, err := h.interactor.RequestWalletConnectionReview(ctx, metadata.TraceID, metadata.Hostname)
 		if err != nil {
-			if errDetails := handleRequestFlowError(ctx, traceID, h.interactor, err); errDetails != nil {
+			if errDetails := handleRequestFlowError(ctx, metadata.TraceID, h.interactor, err); errDetails != nil {
 				return nil, errDetails
 			}
-			h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("reviewing the wallet connection failed: %w", err))
+			h.interactor.NotifyError(ctx, metadata.TraceID, InternalError, fmt.Errorf("reviewing the wallet connection failed: %w", err))
 			return nil, internalError(ErrCouldNotConnectToWallet)
 		}
 
 		a, err := preferences.ParseConnectionApproval(rawApproval)
 		if err != nil {
-			h.interactor.NotifyError(ctx, traceID, UserError, err)
+			h.interactor.NotifyError(ctx, metadata.TraceID, UserError, err)
 			continue
 		}
 		approval = a
@@ -94,75 +82,54 @@ func (h *ClientConnectWallet) Handle(ctx context.Context, rawParams jsonrpc.Para
 			return nil, requestInterruptedError(ErrRequestInterrupted)
 		}
 
-		selectedWallet, err := h.interactor.RequestWalletSelection(ctx, traceID, params.Hostname, availableWallets)
+		selectedWallet, err := h.interactor.RequestWalletSelection(ctx, metadata.TraceID, metadata.Hostname, availableWallets)
 		if err != nil {
-			if errDetails := handleRequestFlowError(ctx, traceID, h.interactor, err); errDetails != nil {
+			if errDetails := handleRequestFlowError(ctx, metadata.TraceID, h.interactor, err); errDetails != nil {
 				return nil, errDetails
 			}
-			h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("requesting the wallet selection failed: %w", err))
+			h.interactor.NotifyError(ctx, metadata.TraceID, InternalError, fmt.Errorf("requesting the wallet selection failed: %w", err))
 			return nil, internalError(ErrCouldNotConnectToWallet)
 		}
 
 		if exist, err := h.walletStore.WalletExists(ctx, selectedWallet.Wallet); err != nil {
-			h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("could not verify the wallet existence: %w", err))
+			h.interactor.NotifyError(ctx, metadata.TraceID, InternalError, fmt.Errorf("could not verify the wallet existence: %w", err))
 			return nil, internalError(ErrCouldNotConnectToWallet)
 		} else if !exist {
-			h.interactor.NotifyError(ctx, traceID, UserError, ErrWalletDoesNotExist)
+			h.interactor.NotifyError(ctx, metadata.TraceID, UserError, ErrWalletDoesNotExist)
 			continue
 		}
 
 		w, err := h.walletStore.GetWallet(ctx, selectedWallet.Wallet, selectedWallet.Passphrase)
 		if err != nil {
 			if errors.Is(err, wallet.ErrWrongPassphrase) {
-				h.interactor.NotifyError(ctx, traceID, UserError, wallet.ErrWrongPassphrase)
+				h.interactor.NotifyError(ctx, metadata.TraceID, UserError, wallet.ErrWrongPassphrase)
 				continue
 			}
-			h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("could not retrieve the wallet: %w", err))
+			h.interactor.NotifyError(ctx, metadata.TraceID, InternalError, fmt.Errorf("could not retrieve the wallet: %w", err))
 			return nil, internalError(ErrCouldNotConnectToWallet)
 		}
 		loadedWallet = w
 		break
 	}
 
-	token, err := h.sessions.ConnectWallet(params.Hostname, loadedWallet)
+	token, err := h.sessions.ConnectWallet(metadata.Hostname, loadedWallet)
 	if err != nil {
-		h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("could not connect to a wallet: %w", err))
+		h.interactor.NotifyError(ctx, metadata.TraceID, InternalError, fmt.Errorf("could not connect to a wallet: %w", err))
 		return nil, internalError(ErrCouldNotConnectToWallet)
 	}
 
-	h.interactor.NotifySuccessfulRequest(ctx, traceID, WalletConnectionSuccessfullyEstablished)
+	h.interactor.NotifySuccessfulRequest(ctx, metadata.TraceID, WalletConnectionSuccessfullyEstablished)
 
 	return ClientConnectWalletResult{
 		Token: token,
 	}, nil
 }
 
-func validateConnectWalletParams(rawParams jsonrpc.Params) (ClientConnectWalletParams, error) {
-	if rawParams == nil {
-		return ClientConnectWalletParams{}, ErrParamsRequired
-	}
-
-	params := ClientConnectWalletParams{}
-	if err := mapstructure.Decode(rawParams, &params); err != nil {
-		return ClientConnectWalletParams{}, ErrParamsDoNotMatch
-	}
-
-	if params.Hostname == "" {
-		return ClientConnectWalletParams{}, ErrHostnameIsRequired
-	}
-
-	return params, nil
-}
-
 func isConnectionRejected(approval preferences.ConnectionApproval) bool {
 	return approval != preferences.ApprovedOnlyThisTime
 }
 
-func NewConnectWallet(
-	walletStore WalletStore,
-	interactor Interactor,
-	sessions *session.Sessions,
-) *ClientConnectWallet {
+func NewConnectWallet(walletStore WalletStore, interactor Interactor, sessions *session.Sessions) *ClientConnectWallet {
 	return &ClientConnectWallet{
 		walletStore: walletStore,
 		interactor:  interactor,